+++ frontend/app/compare/layout.tsx
@@ -49,7 +49,7 @@ export default function CompareLayout({ children }: { children: React.ReactNode 
 
   return (
     <>
-      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }} />
+      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd).replace(/</g, "\\u003c") }} />
       {children}
     </>
   );