+++ frontend/app/smart-money/layout.tsx
@@ -79,7 +79,7 @@ export default function SmartMoneyLayout({ children }: { children: React.ReactNo
 
   return (
     <>
-      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd) }} />
+      <script type="application/ld+json" dangerouslySetInnerHTML={{ __html: JSON.stringify(jsonLd).replace(/</g, "\\u003c") }} />
       {children}
     </>
   );